Extracted base class ExplicitTimeIntegrator from ActuallyExplicitEuler
Refs #11855
